Revert "Move process_outbound_messages to Client"
This reverts commit 498bbd678be281d3e41314c117825089d42e75d2.
