+++ xmtp/src/conversations.rs
@@ -1,9 +1,17 @@
+use std::time::Duration;
+
 use diesel::Connection;
 use futures::executor::block_on;
 use log::info;
 use prost::Message;
 use vodozemac::olm::{self, OlmMessage};
-use xmtp_proto::xmtp::v3::message_contents::{InvitationV1, PadlockMessagePayload};
+use xmtp_proto::xmtp::{
+    message_api::v1::{Envelope, PublishRequest},
+    v3::message_contents::{
+        EdDsaSignature, InvitationV1, PadlockMessageEnvelope, PadlockMessageHeader,
+        PadlockMessagePayload, PadlockMessagePayloadVersion, PadlockMessageSealedMetadata,
+    },
+};
 
 use crate::{
     contact::Contact,
@@ -13,11 +21,12 @@ use crate::{
     session::SessionManager,
     storage::{
         now, ConversationState, DbConnection, InboundInvite, InboundInviteStatus, InboundMessage,
-        InboundMessageStatus, MessageState, NewStoredMessage, RefreshJob, RefreshJobKind,
-        StorageError, StoredConversation, StoredUser,
+        InboundMessageStatus, MessageState, NewStoredMessage, OutboundPayloadState, RefreshJob,
+        RefreshJobKind, StorageError, StoredConversation, StoredMessage, StoredOutboundPayload,
+        StoredSession, StoredUser,
     },
     types::networking::XmtpApiClient,
-    utils::build_installation_message_topic,
+    utils::{base64_encode, build_installation_message_topic},
     vmac_protos::ProtoWrapper,
     Client,
 };
@@ -448,17 +457,191 @@ where
         // Adjust for padding and ensure start_time > 0
         std::cmp::max(job.last_run - PADDING_TIME_NS, 0)
     }
+
+    fn create_outbound_payload(
+        &self,
+        session: &mut SessionManager,
+        message: &StoredMessage,
+    ) -> Result<StoredOutboundPayload, ConversationError> {
+        let is_prekey_message = !session.has_received_message();
+
+        let metadata = PadlockMessageSealedMetadata {
+            sender_user_address: self.client.wallet_address(),
+            sender_installation_id: self.client.account.contact().installation_id(),
+            recipient_user_address: session.user_address(),
+            recipient_installation_id: session.installation_id(),
+            is_prekey_message,
+        };
+        // TODO encrypted sealed metadata using sealed sender
+        let sealed_metadata = metadata.encode_to_vec();
+        let message_header = PadlockMessageHeader {
+            sent_ns: message.created_at as u64,
+            sealed_metadata,
+        };
+        let header_bytes = message_header.encode_to_vec();
+        // TODO expose a vmac method to sign bytes rather than string
+        // https://matrix-org.github.io/vodozemac/vodozemac/olm/struct.Account.html#method.sign
+        let header_signature = self.client.account.sign(&base64_encode(&header_bytes));
+        let header_signature = EdDsaSignature {
+            bytes: header_signature.to_bytes().to_vec(),
+        };
+
+        let payload = PadlockMessagePayload {
+            message_version: PadlockMessagePayloadVersion::One as i32,
+            header_signature: Some(header_signature),
+            convo_id: message.convo_id.clone(),
+            content_bytes: message.content.clone(),
+        };
+        let olm_message = session.encrypt(&payload.encode_to_vec());
+
+        let ciphertext = match olm_message {
+            olm::OlmMessage::Normal(message) => message.to_bytes(),
+            olm::OlmMessage::PreKey(prekey_message) => prekey_message.to_bytes(),
+        };
+        let envelope: PadlockMessageEnvelope = PadlockMessageEnvelope {
+            header_bytes,
+            ciphertext,
+        };
+        Ok(StoredOutboundPayload::new(
+            message.created_at,
+            build_installation_message_topic(&session.installation_id()),
+            envelope.encode_to_vec(),
+            OutboundPayloadState::Pending as i32,
+            0,
+        ))
+    }
+
+    pub async fn process_outbound_message(
+        &self,
+        message: &StoredMessage,
+    ) -> Result<(), ConversationError> {
+        let peer_address =
+            peer_addr_from_convo_id(&message.convo_id, &self.client.wallet_address())?;
+
+        // Refresh remote installations
+        self.client
+            .refresh_user_installations_if_stale(&peer_address)
+            .await?;
+        self.client.store.conn().unwrap().transaction(
+            |transaction| -> Result<(), ConversationError> {
+                let my_sessions = self
+                    .client
+                    .store
+                    .get_latest_sessions(&self.client.wallet_address(), transaction)?;
+                let their_user_addr =
+                    peer_addr_from_convo_id(&message.convo_id, &self.client.wallet_address())?;
+                let their_sessions = self
+                    .client
+                    .store
+                    .get_latest_sessions(&their_user_addr, transaction)?;
+                if their_sessions.is_empty() {
+                    return Err(ConversationError::NoSessions(their_user_addr));
+                }
+
+                let mut outbound_payloads = Vec::new();
+                let mut updated_sessions = Vec::new();
+                for stored_session in my_sessions.iter().chain(&their_sessions) {
+                    if stored_session.peer_installation_id
+                        == self.client.account.contact().installation_id()
+                    {
+                        continue;
+                    }
+                    let mut session = SessionManager::try_from(stored_session)?;
+                    let outbound_payload = self.create_outbound_payload(&mut session, message)?;
+                    let updated_session = StoredSession::try_from(&session)?;
+                    outbound_payloads.push(outbound_payload);
+                    updated_sessions.push(updated_session);
+                }
+
+                self.client.store.commit_outbound_payloads_for_message(
+                    message.id,
+                    MessageState::LocallyCommitted,
+                    outbound_payloads,
+                    updated_sessions,
+                    transaction,
+                )?;
+                Ok(())
+            },
+        )?;
+
+        Ok(())
+    }
+
+    pub async fn process_outbound_messages(&self) -> Result<(), ConversationError> {
+        //Refresh self installations
+        self.client
+            .refresh_user_installations_if_stale(&self.client.wallet_address())
+            .await?;
+        let mut messages = self.client.store.get_unprocessed_messages()?;
+        log::debug!("Processing {} messages", messages.len());
+        messages.sort_by(|a, b| a.created_at.cmp(&b.created_at));
+        for message in messages {
+            if let Err(e) = self.process_outbound_message(&message).await {
+                log::error!(
+                    "Couldn't process message with ID {} because of error: {:?}",
+                    message.id,
+                    e
+                );
+                // TODO update message status to failed on non-retryable errors so that we don't retry it next time
+            }
+        }
+
+        self.publish_outbound_payloads().await?;
+        Ok(())
+    }
+
+    pub async fn publish_outbound_payloads(&self) -> Result<(), ConversationError> {
+        let unsent_payloads = self.client.store.fetch_and_lock_outbound_payloads(
+            OutboundPayloadState::Pending,
+            Duration::from_secs(60).as_nanos() as i64,
+        )?;
+
+        if unsent_payloads.is_empty() {
+            return Ok(());
+        }
+
+        let envelopes = unsent_payloads
+            .iter()
+            .map(|payload| Envelope {
+                content_topic: payload.content_topic.clone(),
+                timestamp_ns: payload.created_at_ns as u64,
+                message: payload.payload.clone(),
+            })
+            .collect();
+
+        // TODO: API tokens
+        self.client
+            .api_client
+            .publish("".to_string(), PublishRequest { envelopes })
+            .await?;
+
+        let payload_ids = unsent_payloads
+            .iter()
+            .map(|payload| payload.created_at_ns)
+            .collect();
+        self.client.store.update_and_unlock_outbound_payloads(
+            payload_ids,
+            OutboundPayloadState::ServerAcknowledged,
+        )?;
+        Ok(())
+    }
 }
 
 #[cfg(test)]
 mod tests {
+    use prost::Message;
     use xmtp_proto::xmtp::message_api::v1::QueryRequest;
 
     use crate::{
+        codecs::{text::TextCodec, ContentCodec},
+        conversation::convo_id,
         conversations::Conversations,
         invitation::Invitation,
         mock_xmtp_api_client::MockXmtpApiClient,
-        storage::{now, InboundInvite, InboundInviteStatus, StoredConversation, StoredUser},
+        storage::{
+            now, InboundInvite, InboundInviteStatus, MessageState, StoredConversation,
+            StoredMessage, StoredUser,
+        },
         test_utils::test_utils::{
             gen_test_client, gen_test_client_internal, gen_test_conversation, gen_two_test_clients,
         },
@@ -495,6 +678,39 @@ mod tests {
     }
 
     #[tokio::test]
+    async fn create_outbound_payload() {
+        let alice_client = gen_test_client().await;
+        let bob_client = gen_test_client().await;
+
+        let conversations = Conversations::new(&alice_client);
+        let mut session = alice_client
+            .get_session(
+                &mut alice_client.store.conn().unwrap(),
+                &bob_client.account.contact(),
+            )
+            .unwrap();
+
+        let _payload = conversations
+            .create_outbound_payload(
+                &mut session,
+                &StoredMessage {
+                    id: 0,
+                    created_at: 0,
+                    convo_id: convo_id(alice_client.wallet_address(), bob_client.wallet_address()),
+                    addr_from: alice_client.wallet_address(),
+                    sent_at_ns: 0,
+                    content: TextCodec::encode("Hello world".to_string())
+                        .unwrap()
+                        .encode_to_vec(),
+                    state: MessageState::Unprocessed as i32,
+                },
+            )
+            .unwrap();
+
+        // TODO validate the payload when implementing the receiver side
+    }
+
+    #[tokio::test]
     async fn process_outbound_messages() {
         let (alice_client, bob_client) = gen_two_test_clients().await;
 
@@ -503,7 +719,7 @@ mod tests {
             gen_test_conversation(&conversations, &bob_client.wallet_address()).await;
 
         conversation.send_text("Hello world").await.unwrap();
-        alice_client.process_outbound_messages().await.unwrap();
+        conversations.process_outbound_messages().await.unwrap();
         let response = bob_client
             .api_client
             .query(QueryRequest {
@@ -643,7 +859,8 @@ mod tests {
         // Send First Message
 
         a_to_b.send_text("Hi").await.unwrap();
-        alice_client.process_outbound_messages().await.unwrap();
+        a_convos.process_outbound_messages().await.unwrap();
+        a_convos.publish_outbound_payloads().await.unwrap();
         b_convos.receive().unwrap();
 
         let bob_messages = bob_client
@@ -681,7 +898,12 @@ mod tests {
             .unwrap();
 
         b_to_a.send_text("Reply").await.unwrap();
-        bob_client.process_outbound_messages().await.unwrap();
+        bob_client
+            .refresh_user_installations(&bob_address)
+            .await
+            .unwrap();
+        b_convos.process_outbound_messages().await.unwrap();
+        b_convos.publish_outbound_payloads().await.unwrap();
 
         a_convos.receive().unwrap();
